requirements.txt
# Keep the versions locked to avoid supply chain attacks 

# PyTorch (CPU-only)
torch==2.9.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# LLM providers
groq==0.32.0
openai==1.58.1  # for OpenAI and Azure OpenAI support

# Vector database and embeddings
chromadb==1.3.4
sentence-transformers==5.1.1

# Web framework
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20  # for file uploads

# Utilities
python-dotenv==1.1.1
pyyaml==6.0.3
pdfplumber==0.11.7
beautifulsoup4==4.14.2
playwright==1.55.0
requests==2.32.3

# spaCy + NLP model
spacy==3.8.9
--extra-index-url https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Testing
pytest==8.4.2
pytest-asyncio==1.2.0

# PDF parsing
pypdf==6.1.1

# Tracing - LangFuse
langfuse>=2.0.0,<3.0.0
pydantic==2.12.0  # for data validation and serialization
